+++ changelog.md
@@ -6,6 +6,7 @@
 * `[usage]` Added github update checker
 * `[usage]` Moved `CommandsColorTag`, `OptionsColorTag`, `Breadcrumbs` to `Info` struct (_incompatible changes_)
 * `[fsutil]` Now `ListingFilter` must be passed as value instead of pointer (_incompatible changes_)
+* `[fsutil]` Added support of filtering by size for `ListingFilter`
 
 ---